Move `stl_reader.h` to `include/stl_reader` and provide CMakeLists.txt
To better support CMake workflows using `FetchContent` and for a better
integration into existing CMake based projects, `stl_reader.h` was moved
to the sub-directory `include/stl_reader`. A `CMakeLists.txt` script is
provided which defines an interface library `stl_reader`.
