+++ docs/doxygen_config/stl_reader.dconf
@@ -907,7 +907,7 @@ WARN_LOGFILE           =
 # spaces. See also FILE_PATTERNS and EXTENSION_MAPPING
 # Note: If this tag is empty the current directory is searched.
 
-INPUT                  = stl_reader.h \
+INPUT                  = include/stl_reader/stl_reader.h \
                          Readme.md
 
 # This tag can be used to specify the character encoding of the source files
